+++ packages/web/src/components/aditude-ad-banner.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { session } from '@lib/session-store'
+  import { session } from '@lib/stores'
   import { afterUpdate, onMount } from 'svelte'
 
   let renderAd = false
