+++ packages/web/src/components/auth-link.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { session } from '@lib/session-store'
+  import { session } from '@lib/stores'
   import Icon from '@components/icon.svelte'
 </script>
