+++ packages/web/src/components/auth-menu.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import UserMenu from '@components/user-menu.svelte'
-  import { session } from '@lib/session-store'
+  import { session } from '@lib/stores'
   import Icon from '@components/icon.svelte'
 </script>
